requirements.txt
requests
2captcha-python
python-dotenv
colorama
pyfiglet